Refactor : QueryHandler 수정 Class 이름 수정
