+++ order-api/src/main/java/org/ecommerce/orderapi/order/service/OrderItemReadService.java
@@ -1,4 +1,4 @@
-package org.ecommerce.orderapi.order.handler;
+package org.ecommerce.orderapi.order.service;
 
 import static org.ecommerce.orderapi.order.util.OrderPolicyConstants.*;
 
@@ -17,7 +17,7 @@ import lombok.RequiredArgsConstructor;
 @Component
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
-public class OrderItemQueryHandler {
+public class OrderItemReadService {
 
 	private final OrderItemRepository orderItemRepository;
